post-management/requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
click==8.1.3
coverage==7.1.0
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
pip==22.0.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2